feat: refactor to new structure
