+++ Makefile
@@ -42,7 +42,7 @@ generate-proto:
 	@echo "Generating proto files for Go"
 	protoc -I=internal/jams-proto/proto/api/v1/ --go_out=clients/go/jams --go-grpc_out=clients/go/jams jams.proto
 	@echo "Generating proto files for Python"
-	python -m grpc_tools.protoc -I=internal/jams-proto/proto/api/v1/ --python_out=clients/python/jams/src/jams/proto --pyi_out=clients/python/jams/src/jams/proto --grpc_python_out=clients/python/jams/src/jams/proto jams.proto
+	python -m grpc_tools.protoc -I=internal/jams-proto/proto/api/v1/ --python_out=clients/python/jams/src/client/types/proto --pyi_out=clients/python/jams/src/client/types/proto --grpc_python_out=clients/python/jams/src/client/types/proto jams.proto
 #	@echo "Generating proto files for TypeScript"
 #	protoc -I=jams-serve/proto/api/v1/ --plugin="$(which protoc-gen-ts)" --ts_proto_opt=esModuleInterop=true --ts_proto_out="clients/typescript/generated" jams.proto
 	@echo "Generating proto files for Java"
