+++ clients/python/jams/src/jams/client/http.py
@@ -1,10 +1,9 @@
 import requests
 
-from src.jams.http import models
-from src.jams.common import type
+from src.jams.client.models import common, http
 
 
-class HttpClient:
+class Client:
     def __init__(self, base_url: str):
         self.base_url = f"http://{base_url}"
 
@@ -17,9 +16,9 @@ class HttpClient:
         except requests.exceptions.ConnectionError:
             raise ConnectionError
 
-    def predict(self, model_name: str, model_input: str) -> type.Prediction:
+    def predict(self, model_name: str, model_input: str) -> common.Prediction:
         url = f"{self.base_url}/api/predict"
-        request = models.PredictRequest(
+        request = http.PredictRequest(
             model_name=model_name, input=model_input
         ).model_dump()
 
@@ -28,8 +27,8 @@ class HttpClient:
             if response.status_code != 200:
                 raise Exception(f"predict failed with {response.status_code}")
             try:
-                resp_obj = models.PredictResponse.model_validate(response.json())
-                return type.Prediction(resp_obj.output)
+                resp_obj = http.PredictResponse.model_validate(response.json())
+                return common.Prediction(resp_obj.output)
             except ValueError as e:
                 raise Exception(f"fail to parse predict response: {e}")
         except requests.exceptions.ConnectionError:
@@ -37,7 +36,7 @@ class HttpClient:
 
     def add_model(self, model_name: str) -> None:
         url = f"{self.base_url}/api/models"
-        request = models.AddModelsRequest(model_name=model_name).model_dump()
+        request = http.AddModelsRequest(model_name=model_name).model_dump()
         try:
             response = requests.post(url, json=request)
             if response.status_code != 200:
@@ -47,7 +46,7 @@ class HttpClient:
 
     def update_model(self, model_name: str) -> None:
         url = f"{self.base_url}/api/models"
-        request = models.UpdateModelsRequest(model_name=model_name).model_dump()
+        request = http.UpdateModelsRequest(model_name=model_name).model_dump()
         try:
             response = requests.put(url, json=request)
             if response.status_code != 200:
@@ -64,7 +63,7 @@ class HttpClient:
         except requests.exceptions.ConnectionError:
             raise ConnectionError
 
-    def get_models(self) -> models.GetModelsResponse:
+    def get_models(self) -> http.GetModelsResponse:
         url = f"{self.base_url}/api/models"
         try:
             response = requests.get(url)
@@ -73,6 +72,6 @@ class HttpClient:
         if response.status_code != 200:
             raise Exception(f"get model failed with {response.status_code}")
         try:
-            return models.GetModelsResponse.model_validate(response.json())
+            return http.GetModelsResponse.model_validate(response.json())
         except ValueError as e:
             raise Exception(f"fail to parse get model response: {e}")
