+++ clients/python/jams/src/jams/client/models/common.py
