+++ clients/python/jams/src/jams/client/models/http.py
