+++ clients/python/jams/src/jams/client/models/proto/jams_pb2_grpc.py
@@ -4,7 +4,8 @@ import grpc
 import warnings
 
 from google.protobuf import empty_pb2 as google_dot_protobuf_dot_empty__pb2
-from src.jams.proto import jams_pb2 as jams__pb2
+
+import src.jams.client.models.proto.jams_pb2 as jams__pb2
 
 GRPC_GENERATED_VERSION = '1.64.1'
 GRPC_VERSION = grpc.__version__
