+++ clients/python/jams/tests/client/grpc/test_health_check.py
@@ -1,4 +1,4 @@
-from src.jams.grpc import client
+from src.jams.client import grpc
 
 from tests.helper import get_grpc_url
 
@@ -6,7 +6,7 @@ from tests.helper import get_grpc_url
 def test_successfully_makes_health_check_request() -> None:
     # Arrange
     base_url = get_grpc_url()
-    grpc_client = client.GrpcClient(base_url)
+    grpc_client = grpc.Client(base_url)
 
     # Act
     grpc_client.health_check()
