+++ clients/python/jams/tests/client/http/test_models.py
@@ -1,5 +1,5 @@
-from src.jams.http import client
-from src.jams.http.models import GetModelsResponse
+from src.jams.client import http
+from src.jams.client.models.http import GetModelsResponse
 
 from tests.helper import get_http_url
 
@@ -7,7 +7,7 @@ from tests.helper import get_http_url
 def test_successfully_makes_get_models_request() -> None:
     # Arrange
     base_url = get_http_url()
-    http_client = client.HttpClient(base_url)
+    http_client = http.Client(base_url)
 
     # Act
     resp = http_client.get_models()
@@ -20,7 +20,7 @@ def test_successfully_makes_get_models_request() -> None:
 def test_successfully_makes_delete_model_request() -> None:
     # Arrange
     base_url = get_http_url()
-    http_client = client.HttpClient(base_url)
+    http_client = http.Client(base_url)
 
     # Act
     http_client.delete_model(model_name="my_awesome_californiahousing_model")
@@ -32,7 +32,7 @@ def test_successfully_makes_delete_model_request() -> None:
 def test_successfully_makes_add_model_request() -> None:
     # Arrange
     base_url = get_http_url()
-    http_client = client.HttpClient(base_url)
+    http_client = http.Client(base_url)
 
     # Act
     http_client.delete_model(model_name="my_awesome_penguin_model")
@@ -45,7 +45,7 @@ def test_successfully_makes_add_model_request() -> None:
 def test_successfully_makes_update_model_request() -> None:
     # Arrange
     base_url = get_http_url()
-    http_client = client.HttpClient(base_url)
+    http_client = http.Client(base_url)
 
     # Act
     http_client.update_model(model_name="titanic_model")
